requirements.txt
Flask==2.2.3
jsonpickle==3.0.1
numpy==1.24.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
paddle==1.0.2
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
Pillow==9.4.0

file = request.files['front']
    try:
        img = Image.open(file.stream)
        img.save(f"uploads/front-{rand_front}.png")
    except:
        return Response(response=jsonpickle.encode({"Error_message": "Couldn't get your Image"}), status=200, mimetype="application/json")
    try:
        read = read_passport(f"uploads/front-{rand_front}.png")
    except:
        return Response(response=jsonpickle.encode({"Error_message": "Couldn't process your image"}), status=200,
                        mimetype="application/json")